+++ src/rules.ts
@@ -16,3 +16,4 @@ import './html/dom-module-name-or-is';
 import './html/move-style-into-template';
 import './html/undefined-elements';
 import './polymer/unbalanced-delimiters';
+import './polymer/set-unknown-attribute';